requirements.txt
langchain==0.3.17
langchain_community==0.3.16
langchain_core==0.3.33
langchain_huggingface==0.1.2
pypdf==5.2.0
python-dotenv==1.0.1
Requests==2.32.3
rich==13.9.4
streamlit==1.31.1
huggingface-hub==0.28.1
